requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
Pillow
python-dotenv 
google-genai